Adds step template for Xamarin Insights dSYM upload
This step template uploads a dSYM symbols file from an iOS app build to Xamarin Insights, an app telemetry services. This allows logged error telemetry to be more readable and understandable with full source file name and location references. The script expects a nuget package with the *.app.dSYM.zip to have been uploaded. It requires an app specific API Key from the service, a step name, and optionally a further search path for specificity.