Don't warn if Restriction already exists
Changed Write-Warning to Write-Host to prevent the step from being
flagged with a warning during the deployment. It will be very common
during deployments (particularly update deployments) that the CGI/ISAPI
Restriction will already exist. This does not seem like an appropriate
scenario for generating a warning during the deployment.