Fixed path issues
The Get-WebConfiguration's filtering with the path attribute was not
reliably finding the $cgiIsapiExtensionPath as a string without escaped
quotes around it. Escaping it in this fashion resulted in the "" around
the path being present in the IIS value. Changed the -Filter to drop the
path attribute and instead loop over all of the "add" elements in order
to find an existing match of a path with the $cgiIsapiExtensionPath .
Additionally, changed this path comparison to use their full paths by
resolving environment variables and relative paths. Not doing so
resulted in no match being found between ultimately equivalent paths and
Add-WebConfiguration would fail due to a duplicate path attempting to be
added.