Add support for other DacFX commands
Can now generate deployment reports and deployment scripts which are loaded back into Octopus Deploy as artefacts.  This gives the option of creating a multi step process with manual intervention in the middle.

Dropped support of VS version of SSDT.  The script now searches %Program Files% looking for the most recent version of DacFX deployed on the server.