Create Wait-until-time
This step template will pause the octopus process until a given time.

Intended use:
Any time when the process needs to be halted. For example when using a Manual Intervention step to approve the deploy but the user wants the deploy process to continue automatically at night.

Description
It has 2 parameters: DefaultTargetTime that is a step parameter and TargetTime that is a release variable.
If no configuration is given (DefaultTargetTime and TargetTime is not set) the step will continue emmidiately.
If the step parameter DefaultTargetTime is set but not TargetTime, the step will wait until DefaultTargetTime
If TargetTime is set in the release variables for the environment that the release is targeting the step will wait until TargetTime. 
It is possible to make TargetTime as a promptable variable to force the user to enter time every time the release is deployed.