Added support for assigning user to multiple roles
We have many apps where we prefer to assign a few different roles (e.g. `db_datareader` and `db_datawriter`) so I propose an update which allows multiple roles if comma separated. Also I've uncommented the `Exit 1` statement in the catch block, the script should fail the deployment in my opinion if it can't create the user or login otherwise you end up with a successful deployment which didn't actually succeed in one of the key steps.