Added Import-Module statement
Script assumes that since the module is installed, it is available.  I found on many of my servers that the path to the DLM modules were formatted incorrectly and though the modules were installed, they were not available.  You can run an Import-Module statement to correct this, which is what I've added