Update step-templates/ssis-deploy-ispac-from-package.json
Added the management of the package variables with the same name. When a package has the same variable name of the others dtsx, the step template will deploy just to one var, binding it to each package (with a single value). In our case, a variable like "OffsetInDays" should be evaluate with different values, one per package. So I've added the management, but the naming convention of the variables to bind has changed. The new spec follows the rule PackageNameWithoutExtension.VariableName