depth increase from powershells default of 2
I've added a "depth 99" to the json convert as without it, files that have more than 2 levels will not be converted back properly when saved. This does mean that if someone has a file with 100 levels it will mess up the 100th level but I highly doubt anyone would make a config file for that.