Create octopus-logoctopusparameters.json
octopus-LogOctopusParameters creates and writes a powershell script to the Release Task Log during deployment. This powershell script can be copy/pasted to a powershell console, and it will generate an $OctopusParameters variable with all of the values used during the deployment, along with helper methods .ToGridView(), .ToScreen(), and .ToClipBoard() which should be self explanatory. This should help with debugging deployments.