Update Powershell script to allow for regex tags in Channel Rules
I ran into an error with deployments in my environment due to the fact that the $_.Tag is expected to be a string only. Wrapping it in the URLEncode seems to have fixed the issue and allows the script to evaluate regex expressions (which are allowed in Channel Version Rules).