Null tenant list issues
If tenanted only deployments is enabled and no tenants are passed in I've added a check and fail if this is the case. I've also added a check to see if the tenant list is null before it checks to finding match tenants. This is to fix an issue where by if no tenants are in the space then the script would bring all tenants back and select the first tenant in the list and associate that deployment target with the tenant.