Update iis-website-delete.json
Added a filter to exclude the default SSL-binding from being deleted. This will break all sites relying on this certificate if the site is not configured for SNI (checkbox in SSL) because they too will loose the reference to the binding. Using Remove-Website will remove the appropriate SSL-binding reference, so skipping is enough in this scenario. All dedicated (named) bindings will still be removed correctly.