Loosen CSP headers to keep site working
This backs off the previously introduced CSP header and makes the values permissive 
enough to work with the current application code and page structure of the Library 
app. However, usage of `'unsafe-inline'` and `data:` are explicitly discouraged by the CSP
spec, because they are so permissive that it's like not having a CSP header at all.

We should do further analysis of the changes required to tighten up the CSP header
at a later date.

Also fixes the `Permissions-Policy` header to be in the right format, and disable almost
all known browser permissions - none should be relevant for our very simple Library app.