Update NewRelic Deployment Marker step template with polling checks
Right now, there is a chicken and the egg problem with the newrelic deployment marker step template.  It requires the application id.  In order to get the application id you must first deploy the app and have it reporting to newrelic (so it created the app).  This makes is very annoying and painful when adding a new application as your first deploys will always fail.  This change adds a configurable polling functionality to wait till the app shows up and bomb out if the app doesn't appear.