Update aws-add-remove-elbv2.json
When on Linux arm64 and installing the aws module it uses AWSPowerShell.NetCore, but on Windows with the older version it could be AWSPowerShell. So this checks to see if either version was automatically imported.

If it's already imported then it carries on, otherwise it attempts to see if it's installed and then import it. Or throw a message to say either version cannot be found.

```
if (Get-Module | Where-Object { $_.Name -like "AWSPowerShel*" }) {
	Write-Host "AWS PowerShell module is already loaded."
} else {
	if (!(Get-Module -ListAvailable | Where-Object {  $_.Name -like "AWSPowerShel*" })) {
    	Write-Host "AWSPowerShell / AWSPowerShell.NetCore not found"
    } else {
    	Get-Module -ListAvailable | Where-Object {  $_.Name -like "AWSPowerShel*" } | Import-Module
    }
}
```