Deploy Child Project - Add Channel Release Filter
When the step needs to pull back all the releases for a specific channel (for example, when deploying to the first environment in a lifecycle), it will now send in the version number for the server to search on.  This will allow the server to do some filtering prior to sending back the matching releases.  I've also upped the number of releases returned from the default 30 to100.