Update Jenkins Step Template
This update adds additional options when calling the Jenkins /API endpoint that can be used to rate limit the Invoke-RestMethod calls. These calls fetch the URL for the build and depending on the Jenkins server load, may cause excessive consecutive calls while trying to get the build URL. There is also an added dimension of sleeping between calls.